+++ vessel-runtime/src/test/java/com/textnow/android/vessel/VesselCacheUsageTest.kt
@@ -1,5 +1,3 @@
-package com.textnow.android.vessel
-
 /**
  * MIT License
  *
@@ -23,6 +21,8 @@ package com.textnow.android.vessel
  * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
  * SOFTWARE.
  */
+
+package com.textnow.android.vessel
 
 import android.os.Build
 import androidx.arch.core.executor.testing.InstantTaskExecutorRule
@@ -163,6 +163,7 @@ abstract class BaseVesselCacheUsageTest(async: Boolean) : BaseVesselTest<TestCac
         assertThat(profData?.hitCountOf(Span.DELETE_FROM_DB)).isEqualTo(1)
         assertThat(profData?.hitCountOf(Event.CACHE_HIT_READ)).isEqualTo(2)
     }
+
 }
 
 @Config(sdk = [Build.VERSION_CODES.P])