+++ vessel-runtime/src/test/java/com/textnow/android/vessel/VesselWrapper.kt
@@ -1,3 +1,27 @@
+/**
+ * MIT License
+ *
+ * Copyright (c) 2020 TextNow, Inc.
+ *
+ * Permission is hereby granted, free of charge, to any person obtaining a copy
+ * of this software and associated documentation files (the "Software"), to deal
+ * in the Software without restriction, including without limitation the rights
+ * to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+ * copies of the Software, and to permit persons to whom the Software is
+ * furnished to do so, subject to the following conditions:
+ *
+ * The above copyright notice and this permission notice shall be included in all
+ * copies or substantial portions of the Software.
+ *
+ * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+ * IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+ * FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+ * AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+ * LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+ * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
+ * SOFTWARE.
+ */
+
 package com.textnow.android.vessel
 
 import kotlin.reflect.KClass
@@ -7,7 +31,7 @@ import kotlin.reflect.KClass
  * code to test the blocking and suspend functions
  */
 class VesselWrapper(private val vessel: Vessel, private val async: Boolean = false) {
-    suspend fun <T: Any> get(type: KClass<T>): T? {
+    suspend fun <T : Any> get(type: KClass<T>): T? {
         return if (async) {
             vessel.get(type)
         } else {
@@ -23,7 +47,7 @@ class VesselWrapper(private val vessel: Vessel, private val async: Boolean = fal
         }
     }
 
-    suspend fun <T: Any> delete(type: KClass<T>) {
+    suspend fun <T : Any> delete(type: KClass<T>) {
         if (async) {
             vessel.delete(type)
         } else {
@@ -31,12 +55,10 @@ class VesselWrapper(private val vessel: Vessel, private val async: Boolean = fal
         }
     }
 
-    suspend fun preload(timeoutMS: Int?) {
-        if (async) {
-            vessel.preload(timeoutMS)
-        } else {
-            vessel.preloadBlocking(timeoutMS)
-        }
+    suspend fun preload(timeoutMS: Int?) = if (async) {
+        vessel.preload(timeoutMS)
+    } else {
+        vessel.preloadBlocking(timeoutMS)
     }
 
     val profileData get() = vessel.profileData